% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\docType{data}
\name{contour}
\alias{contour}
\title{Moosilauke contour segment plot information to link to \code{location} data 
by \code{PID}.}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 512 rows and 4 columns.
\itemize{
  \item PID: Plot ID to link to \code{location} data in this package.
  \item CONTNAM: Contour name
  \item STPACE: Starting pace of contour plot.
  \item LABEL: More information about a plot location.  Mostly just relevant
for plots in the 'OTHER' category.
}}
\usage{
contour
}
\description{
Moosilauke contour segment plot information to link to \code{location} data 
by \code{PID}.
}
\keyword{datasets}

