% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{make_lines}
\alias{make_lines}
\title{Convert contour data to SpatialLines}
\usage{
make_lines(data, type = c("CONT", "TRAN", "PERM"))
}
\arguments{
\item{data}{Input data}

\item{type}{Type of plot ("CONT"=contour, "TRAN"=transec, "PERM"=permanent)}
}
\description{
Only supports making SpatialLines for contour plots and transects.
}

