% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moosegps.R
\name{moosegps}
\alias{moosegps}
\title{Packaged data for Moosilauke plot locations}
\description{
This includes location information
for permanent plots, transects, and contours (segments), as well as some other
assorted locations referred to as 'OTHER'. Cleaning scripts and raw-data are 
included in directories 'clean' and 'raw'.
}

