% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\docType{data}
\name{other}
\alias{other}
\title{Moosilauke or nearby plots information in the 'OTHER' category.}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 112 rows and 2 columns.
\itemize{
  \item PID: Plot ID to link to other datasets in this package.
  \item LABEL: More information about a plot location.  Mostly just relevant
for plots in the 'OTHER' category.
}}
\usage{
other
}
\description{
Moosilauke or nearby plots information in the 'OTHER' category.
}
\keyword{datasets}

