% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\docType{data}
\name{transect}
\alias{transect}
\title{Moosilauke transect plot specific information.}
\format{An object of class \code{data.table} (inherits from
  \code{data.frame}) with 237 rows and 4 columns.
\itemize{
  \item PID: Plot ID to link to other datasets in this package.
  \item TRAN: Transect id
  \item TPLOT: Transect plot number
  \item LABEL: More information about a plot location.  Mostly just relevant
for plots in the 'OTHER' category.
}}
\usage{
transect
}
\description{
Moosilauke transect plot specific information.
}
\keyword{datasets}

